services/ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
openai==1.59.7
PyJWT==2.10.1
httpx==0.28.1
